requirements.txt
google-cloud-dlp>=3.12.1
llama-index
